requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
pyodbc
python-dotenv
pydantic
pydantic-settings
